backend/evolve_logix/requirements.txt
django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.2.0
djangorestframework-simplejwt==4.5.0
PyJWT==v1.7.1
psycopg2-binary==2.9.9
